chore: fix api

+++ src/main/java/com/example/demo/service/ProductService.java
@@ -39,7 +39,7 @@ public interface ProductService {
 
 	List<Product> getProductByComment();
 
-	List<Product> getProductByFilter(String CateName, String material, String color, String size, String order, Double min, Double max, Integer rating);
+	List<Product> getProductByFilter(String CateName, String material, String color, String size, String order, Double min, Double max, Double rating);
 
 	List<Product> getProductByBuyQuantity();
